Initial batch quad rendering shape done

+++ src/Engine/Renderer/Primitives/QuadMesh.h
@@ -15,17 +15,17 @@ namespace carrot::renderer {
     {
         {
             // top-left
-            { .x = 0.0f, .y = 0.0f, .u = 0.0f, .v = 0.0f, .color = 0xFFFFFFFFu },
+            { .x = 0.f, .y = 0.f, .u = 0.f, .v = 0.f, .color = 0xFFFFFFFFu },
             // top-right
-            { .x = 1.0f, .y = 0.0f, .u = 1.0f, .v = 0.0f, .color = 0xFFFFFFFFu },
+            { .x = 1.f, .y = 0.f, .u = 1.f, .v = 0.f, .color = 0xFFFFFFFFu },
             // bottom-right
-            { .x = 1.0f, .y = 1.0f, .u = 1.0f, .v = 1.0f, .color = 0xFFFFFFFFu },
+            { .x = 1.f, .y = 1.f, .u = 1.f, .v = 1.f, .color = 0xFFFFFFFFu },
             // bottom-left
-            { .x = 0.0f, .y = 1.0f, .u = 0.0f, .v = 1.0f, .color = 0xFFFFFFFFu },
+            { .x = 0.f, .y = 1.f, .u = 0.f, .v = 1.f, .color = 0xFFFFFFFFu },
         }
     };
 
-    inline constexpr std::array<uint16_t, 6> k_unit_quad_indices
+    inline constexpr std::array<uint32_t, 6> k_unit_quad_indices
     {
         {
             0, 1, 2,
